js client point queries now work

+++ drivers/javascript/rethinkdb/query/table.js
@@ -132,10 +132,43 @@ rethinkdb.query.DeleteQuery.prototype.buildQuery = function() {
 };
 
 /**
+ * @param {rethinkdb.query.Table} table
+ * @param {rethinkdb.query.Expression} key
+ * @constructor
+ * @extends {rethinkdb.query.BaseQuery}
+ */
+rethinkdb.query.PointDeleteQuery = function(table, key) {
+    this.table_ = table;
+    this.key_ = key;
+};
+goog.inherits(rethinkdb.query.PointDeleteQuery, rethinkdb.query.BaseQuery);
+
+rethinkdb.query.PointDeleteQuery.prototype.buildQuery = function() {
+    var pointdelete = new WriteQuery.PointDelete();
+    pointdelete.setTableRef(this.table_.compile().getTableOrDefault().getTableRefOrDefault());
+    pointdelete.setAttrname('id');
+    pointdelete.setKey(this.key_.compile());
+
+    var write = new WriteQuery();
+    write.setType(WriteQuery.WriteQueryType.POINTDELETE);
+    write.setPointDelete(pointdelete);
+
+    var query = new Query();
+    query.setType(Query.QueryType.WRITE);
+    query.setWriteQuery(write);
+
+    return query;
+};
+
+/**
  * Deletes all rows the current view
  */
 rethinkdb.query.Expression.prototype.del = function() {
-    return new rethinkdb.query.DeleteQuery(this);
+    if (this instanceof rethinkdb.query.GetExpression) {
+        return new rethinkdb.query.PointDeleteQuery(this.table_, this.key_);
+    } else {
+        return new rethinkdb.query.DeleteQuery(this);
+    }
 };
 goog.exportProperty(rethinkdb.query.Expression.prototype, 'del',
                     rethinkdb.query.Expression.prototype.del);
@@ -173,11 +206,51 @@ rethinkdb.query.UpdateQuery.prototype.buildQuery = function() {
 };
 
 /**
+ * @param {rethinkdb.query.Table} table
+ * @param {rethinkdb.query.Expression} key
+ * @param {rethinkdb.query.FunctionExpression} mapping
+ * @constructor
+ * @extends {rethinkdb.query.BaseQuery}
+ */
+rethinkdb.query.PointUpdateQuery = function(table, key, mapping) {
+    this.table_ = table;
+    this.key_ = key;
+    this.mapping_ = mapping;
+};
+goog.inherits(rethinkdb.query.PointUpdateQuery, rethinkdb.query.BaseQuery);
+
+rethinkdb.query.PointUpdateQuery.prototype.buildQuery = function() {
+    var mapping = new Mapping();
+    mapping.setArg(this.mapping_.args[0]);
+    mapping.setBody(this.mapping_.body.compile());
+
+    var pointupdate = new WriteQuery.PointUpdate();
+    pointupdate.setTableRef(this.table_.compile().getTableOrDefault().getTableRefOrDefault());
+    pointupdate.setAttrname('id');
+    pointupdate.setKey(this.key_.compile());
+    pointupdate.setMapping(mapping);
+
+    var write = new WriteQuery();
+    write.setType(WriteQuery.WriteQueryType.POINTUPDATE);
+    write.setPointUpdate(pointupdate);
+
+    var query = new Query();
+    query.setType(Query.QueryType.WRITE);
+    query.setWriteQuery(write);
+
+    return query;
+};
+
+/**
  * Updates all rows according to the given mapping function
  */
 rethinkdb.query.Expression.prototype.update = function(mapping) {
     mapping = functionWrap_(mapping);
-    return new rethinkdb.query.UpdateQuery(this, mapping);
+    if (this instanceof rethinkdb.query.GetExpression) {
+        return new rethinkdb.query.PointUpdateQuery(this.table_, this.key_, mapping);
+    } else {
+        return new rethinkdb.query.UpdateQuery(this, mapping);
+    }
 };
 goog.exportProperty(rethinkdb.query.Expression.prototype, 'update',
                     rethinkdb.query.Expression.prototype.update);
@@ -215,11 +288,51 @@ rethinkdb.query.MutateQuery.prototype.buildQuery = function() {
 };
 
 /**
+ * @param {rethinkdb.query.Table} table
+ * @param {rethinkdb.query.Expression} key
+ * @param {rethinkdb.query.FunctionExpression} mapping
+ * @constructor
+ * @extends {rethinkdb.query.BaseQuery}
+ */
+rethinkdb.query.PointMutateQuery = function(table, key, mapping) {
+    this.table_ = table;
+    this.key_ = key;
+    this.mapping_ = mapping;
+};
+goog.inherits(rethinkdb.query.PointMutateQuery, rethinkdb.query.BaseQuery);
+
+rethinkdb.query.PointMutateQuery.prototype.buildQuery = function() {
+    var mapping = new Mapping();
+    mapping.setArg(this.mapping_.args[0]);
+    mapping.setBody(this.mapping_.body.compile());
+
+    var pointmutate = new WriteQuery.PointMutate();
+    pointmutate.setTableRef(this.table_.compile().getTableOrDefault().getTableRefOrDefault());
+    pointmutate.setAttrname('id');
+    pointmutate.setKey(this.key_.compile());
+    pointmutate.setMapping(mapping);
+
+    var write = new WriteQuery();
+    write.setType(WriteQuery.WriteQueryType.POINTMUTATE);
+    write.setPointMutate(pointmutate);
+
+    var query = new Query();
+    query.setType(Query.QueryType.WRITE);
+    query.setWriteQuery(write);
+
+    return query;
+};
+
+/**
  * Replcaces each row the the result of the mapping expression
  */
 rethinkdb.query.Expression.prototype.mutate = function(mapping) {
     mapping = functionWrap_(mapping);
-    return new rethinkdb.query.MutateQuery(this, mapping);
+    if (this instanceof rethinkdb.query.GetExpression) {
+        return new rethinkdb.query.PointMutateQuery(this.table_, this.key_, mapping);
+    } else {
+        return new rethinkdb.query.MutateQuery(this, mapping);
+    }
 };
 goog.exportProperty(rethinkdb.query.Expression.prototype, 'mutate',
                     rethinkdb.query.Expression.prototype.mutate);
